An O(n) solution.
Used a HashMap because I can search for an element in O(1).
The idea is that we defined the Hashmap as <value, index in nums>
So for each element we want to see if there is an element with the
the value target - nums[i]. if so you can return the indexes.